update(cherryusb): update to v1.4.0

+++ components/drivers/usb/cherryusb/demo/cdc_ecm_template.c
@@ -269,7 +269,7 @@ struct usbd_interface intf1;
  * sudo ifconfig enxaabbccddeeff up
  * sudo dhcpclient enxaabbccddeeff
 */
-void cdc_ecm_init(uint8_t busid, uint32_t reg_base)
+void cdc_ecm_init(uint8_t busid, uintptr_t reg_base)
 {
     cdc_ecm_lwip_init();
